recommender/requirements.txt
Pillow
numpy
pandas
# tensorflow-gpu==1.12.0
tensorflow-gpu
keras
scikit-learn
matplotlib
sortedcontainers